Move string related funcions to c10/util/string_utils.h.
